requirements.txt
aihub-client==0.6.3
fastapi==0.115.8
uvicorn[standard]==0.34.0
py-ulid==1.0.3
starlette-context==0.3.6
pymongo==4.11.1
requests==2.32.3
python-jose==3.4.0
pydantic-settings==2.7.1
httpx==0.28.1
motor==3.7.0
pytz==2025.1
dependency-injector==4.46.0
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
colorama==0.4.6
deprecated==1.2.18
googleapis-common-protos==1.69.0
grpcio==1.67.1
h11==0.14.0
httptools==0.6.4
idna==3.10
importlib-metadata==8.5.0
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
packaging==24.2
protobuf==5.29.3
pydantic-core==2.27.2
pydantic==2.10.6
python-dotenv==1.0.1
pyyaml==6.0.2
sniffio==1.3.1
starlette==0.45.3
typing-extensions==4.12.2
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0
wrapt==1.17.2
zipp==3.21.0
pymilvus==2.5.5
openai==0.27.10
transformers==4.31.0
langchain~=0.3.17
langchain-community~=0.3.17
langchain_milvus==0.1.8
pymupdf4llm==0.0.24